ログイン中に遷移してしまうバグなどを対処した
