+++ src/components/introduction/index.js
@@ -1,70 +1,36 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React, { useState, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
+import LoadingOverlay from 'react-loading-overlay';
 
-import Header from './Header.js';
-import Footer from './Footer.js';
-import Auth from '../Auth.js';
+import firebase from '../../firebase';
 
-import { makeStyles } from '@material-ui/styles';
-import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
-import Container from '@material-ui/core/Container';
-import Button from '@material-ui/core/Button';
+import Content from './Content';
 
-const styles = makeStyles({
-  top: {
-    backgroundColor: '#33DFEA',
-    margin: '0px',
-    border: 'opx',
-  },
-  container: {
-    'max-width': '740px',
-    color: '#FFFFFF',
-  },
-  topText: {
-    'text-align': 'center',
-    margin: '30px 0px',
-    'font-weight': '300',
-  },
-  button: {
-    left: '38%',
-    'padding' : '5px 10px',
-    'margin' : '30px',
-    color: '#FFFFFF',
-    background: '#EDC124',
-    'text-align': 'center',
-  },
-  bottom: {
-    backgroundColor: '#FFFFFF',
-  }
-});
+export default function Introduction(props) {
+  const history = useHistory();
+  //const classes = styles();
 
-export default function Introduction() {
-  const classes = styles();
-  return (
-    <Auth>
-      <Box>
-        <Box className={classes.top}>
-          <Header />
-          <Container className={classes.container}>
-            <Grid container direction="column">
-              <Grid item>
-                <h1 className={classes.topText}>
-                  大学の課題と<br />
-                  楽しく付き合おう<br />
-                  そんな手助けをするアプリ
-                </h1>
-              </Grid>
-              <Grid item>
-                <Button variant="contained" className={classes.button}>始めてみる</Button>
-              </Grid>
-            </Grid>
-          </Container>
-        </Box>
-        <Box className={classes.bottom}>
-          <Footer />
-        </Box>
-      </Box>
-    </Auth>
+  const [user, setUser] = useState(null);
+
+  useEffect(() => {
+    firebase.auth().onAuthStateChanged(user => {
+      if (user) {
+        history.push("/main")
+      }
+      setUser(user);
+    });
+
+  });
+
+  function Login() {
+    const provider = new firebase.auth.GoogleAuthProvider();
+    firebase.auth().signInWithPopup(provider);
+  };
+
+  return(
+    <Content
+      Login={Login}
+      >
+    </Content>
   );
 }
